+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed data integrity context to `https://w3id.org/security/data-integrity/v2`.
 - Use `eddsa-jcs-2022` cryptosuite when generating minisign identity proofs.
 - Don't reject remote polls without end date.
+- Improved documentation and error messages for `update-config` command.
 
 ### Fixed